fixed some errors?
