+++ web-ui/src/pages/OnboardProgressPage.jsx
@@ -3,37 +3,16 @@ import { DataGrid, GridToolbar } from "@mui/x-data-grid";
 import { Link, useHistory } from "react-router-dom";
 import "./OnboardProgressPage.css";
 import { Box } from "@mui/system";
-import FileUploadIcon from "@mui/icons-material/FileUpload";
 import SearchIcon from "@mui/icons-material/Search";
 import InputAdornment from "@mui/material/InputAdornment";
 import TextField from "@mui/material/TextField";
 import PersonAddIcon from "@mui/icons-material/PersonAdd";
+import AddOnboardeeModal from "../components/modal/AddOnboardeeModal";
+import { AppContext } from "../../../context/AppContext";
+import { createOnboardee } from "../api/onboardeeMember";
 
-import {
-  Button,
-  Modal,
-  Typography,
-  Grid,
-  Divider,
-  IconButton,
-  Autocomplete,
-} from "@mui/material";
+import { Button, Modal, Typography } from "@mui/material";
 import { useState } from "react";
-
-const modalBoxStyle = {
-  position: "absolute",
-  top: "50%",
-  left: "50%",
-  transform: "translate(-50%, -50%)",
-  width: "75%",
-  backgroundColor: "#fff",
-  border: "2px solid #000",
-  boxShadow: 24,
-  pt: 2,
-  px: 4,
-  pb: 3,
-  m: 2,
-};
 
 const modalBoxStyleMini = {
   position: "absolute",
@@ -51,34 +30,22 @@ const modalBoxStyleMini = {
 };
 
 export default function OnboardProgressPage() {
+  const { state, dispatch } = useContext(AppContext);
+  const { csrf } = state;
   const [open, setOpen] = useState(false);
-  const [empFile, setEmpFile] = useState(" ");
-  const [offer, setOfferFile] = useState(" ");
-  const [addOnboardeeModal, setAddOnboardeeModal] = useState(false);
+  const [AddOnboardeeModalBool, setAddOnboardeeModalBool] = useState(false);
 
   const handleOpen = () => setOpen(true);
   const handleClose = () => {
     setOpen(false);
-    setEmpFile(" ");
-    setOfferFile(" ");
   };
   const handleSubmitClose = () => {
     setOpen(false);
-    setAddOnboardeeModal(true);
+    setAddOnboardeeModalBool(true);
   };
   const handleMsgModalClose = () => {
-    setAddOnboardeeModal(false);
+    setAddOnboardeeModalBool(false);
   };
-  const handleEmployeeAgreement = (e) => {
-    setEmpFile(e.target.value.replace(/^.*[\\/]/, ""));
-  };
-  const handleOfferLetter = (e) => {
-    setOfferFile(e.target.value.replace(/^.*[\\/]/, ""));
-  };
-
-  const posOptions = ["dummy1", "dummy2", "dummy3"];
-  const hireOptions = ["dummy4", "dummy5", "dummy6"];
-  const pdlOptions = ["dummy7", "dummy8", "dummy9"];
 
   const history = useHistory();
   const handleRowClick = (name, email, hireType, userID) => {
@@ -208,207 +175,34 @@ export default function OnboardProgressPage() {
           onboardee={{}}
           open={open}
           onClose={handleClose}
-          onSave={async (member) => {
+          onSave={async (onboardee) => {
             if (
               onboardee.employeeId &&
               onboardee.firstName &&
               onboardee.lastName &&
-              monboardee.position &&
-              member.email &&
-              member.hireType &&
-              member.pdl &&
+              onboardee.position &&
+              onboardee.email &&
+              onboardee.hireType &&
+              onboardee.pdl &&
               csrf
             ) {
-              let res = await createMember(member, csrf);
+              let res = await createOnboardee(onboardee, csrf);
               let data =
                 res.payload && res.payload.data && !res.error
                   ? res.payload.data
                   : null;
               if (data) {
                 dispatch({
-                  type: UPDATE_MEMBER_PROFILES,
-                  payload: [...memberProfiles, data],
+                  type: UPDATE_ONBOARDEE_PROFILES,
+                  payload: [...onboardeeProfiles, data],
                 });
               }
-              handleClose();
+              handleSubmitClose();
             }
           }}
         />
-        {/* <Modal
-          open={open}
-          onClose={handleClose}
-          aria-labelledby="title"
-          aria-describedby="description"
-        > */}
-        {/* <Box sx={modalBoxStyle}>
-            <Typography align="center" id="title" variant="h3" component="h2">
-              Add Onboardee:
-            </Typography>
-            <Grid container space={2}>
-              <Grid item xs={6}>
-                <Typography id="description" sx={{ mt: 2 }}>
-                  Position:
-                </Typography>
-                <Autocomplete
-                  disablePortal
-                  options={posOptions}
-                  sx={{ width: "75%" }}
-                  renderInput={(option) => (
-                    <TextField variant="outlined" {...option} />
-                  )}
-                />
-              </Grid>
-              <Grid item xs={6}>
-                <Typography id="description" sx={{ mt: 2 }}>
-                  Hire Type:
-                </Typography>
-                <Autocomplete
-                  disablePortal
-                  options={hireOptions}
-                  sx={{ width: "75%" }}
-                  renderInput={(option) => (
-                    <TextField variant="outlined" {...option} />
-                  )}
-                />
-              </Grid>
-            </Grid>
-            <Grid container space={2}>
-              <Grid item xs={6}>
-                <Typography id="description" sx={{ mt: 2 }}>
-                  First Name:
-                </Typography>
-                <TextField
-                  sx={{ width: "75%" }}
-                  id="firstName"
-                  variant="outlined"
-                />
-              </Grid>
-              <Grid item xs={6}>
-                <Typography id="description" sx={{ mt: 2 }}>
-                  Last Name:
-                </Typography>
-                <TextField
-                  sx={{ width: "75%" }}
-                  id="lastName"
-                  variant="outlined"
-                />
-              </Grid>
-            </Grid>
-            <Grid container space={3}>
-              <Grid item xs={6}>
-                <Typography id="description" sx={{ mt: 2 }}>
-                  Email:
-                </Typography>
-                <TextField
-                  sx={{ width: "75%" }}
-                  id="email"
-                  variant="outlined"
-                />
-              </Grid>
-              <Grid item xs={6}>
-                <Typography id="description" sx={{ mt: 2 }}>
-                  PDL/Manager:
-                </Typography>
-                <Autocomplete
-                  disablePortal
-                  options={pdlOptions}
-                  sx={{ width: "75%" }}
-                  renderInput={(option) => (
-                    <TextField variant="outlined" {...option} />
-                  )}
-                />
-              </Grid>
-            </Grid>
-            <Divider sx={{ m: 3 }} variant="middle" />
-            <Grid
-              container
-              flexDirection="column"
-              alignItems="center"
-              rowSpacing={3}
-            >
-              <Grid item xs={"auto"}>
-                <Typography
-                  align="center"
-                  id="description"
-                  sx={{ mt: 0, display: "inline-flex" }}
-                >
-                  Offer Letter:
-                </Typography>
-                <IconButton component="label">
-                  <input
-                    hidden
-                    accept=".pdf"
-                    type="file"
-                    id="offerLetter"
-                    onChange={handleOfferLetter}
-                  />
-                  <FileUploadIcon />
-                </IconButton>
-                <Typography
-                  sx={{
-                    display: "inline-flex",
-                    fontStyle: "italic",
-                    fontSize: 12,
-                    marginLeft: 5,
-                  }}
-                >
-                  {offer}
-                </Typography>
-              </Grid>
-              <Grid item id="description" xs={"auto"}>
-                <Typography
-                  align="center"
-                  id="description"
-                  sx={{ mt: 2, display: "inline-flex" }}
-                >
-                  Employment Agreement:
-                </Typography>
-                <IconButton component="label">
-                  <input
-                    hidden
-                    accept=".pdf"
-                    type="file"
-                    id="empAgreement"
-                    onChange={handleEmployeeAgreement}
-                  />
-                  <FileUploadIcon />
-                </IconButton>
-                <Typography
-                  sx={{
-                    display: "inline-flex",
-                    fontStyle: "italic",
-                    fontSize: 12,
-                    marginLeft: 5,
-                  }}
-                >
-                  {empFile}
-                </Typography>
-              </Grid>
-            </Grid>
-            <Grid container>
-              <Grid
-                item
-                xs={6}
-                style={{ display: "flex", justifyContent: "flex-start" }}
-              >
-                <Button variant="contained" onClick={handleClose}>
-                  Cancel
-                </Button>
-              </Grid>
-              <Grid
-                item
-                xs={6}
-                style={{ display: "flex", justifyContent: "flex-end" }}
-              >
-                <Button variant="contained" onClick={handleSubmitClose}>
-                  Submit
-                </Button>
-              </Grid>
-            </Grid>
-          </Box>
-        </Modal> */}
         <Modal
-          open={AddOnboardeeModal}
+          open={AddOnboardeeModalBool}
           onClose={handleClose}
           aria-labelledby="title"
           aria-describedby="description"
